--- application/public/robots.txt
@@ -1,4 +1,0 @@
-User-agent: *
-Allow: /
-
-Sitemap: https://www.damianczerwinski.pl/keep-sticky-notes/sitemap.xml
