+++ application/src/components/Footer.js
@@ -1,52 +1,16 @@
-import React from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
-import { faSun, faMoon } from '@fortawesome/free-solid-svg-icons';
-import { useTheme } from '../hooks/useTheme';
-
 export const Footer = () => {
-    const { theme, toggleTheme } = useTheme();
+    const currentYear = new Date().getFullYear();
 
     return (
         <footer className="footer">
             <div className="footer-content">
-                <div className="social-links">
-                    <a
-                        href="https://github.com/damianczer"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="social-link github"
-                        aria-label="GitHub Profile"
-                    >
-                        <FontAwesomeIcon icon={faGithub} />
-                    </a>
-                    <a
-                        href="https://www.linkedin.com/in/daczerw/"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="social-link linkedin"
-                        aria-label="LinkedIn Profile"
-                    >
-                        <FontAwesomeIcon icon={faLinkedin} />
-                    </a>
-                </div>
                 <div className="copyright">
-                    <p>Copyright © 2025 Damian Czerwiński - All Rights Reserved.</p>
+                    <p>Copyright © {currentYear} Keep Sticky Notes - All Rights Reserved.</p>
+                    <p>Created by <a href="https://www.damianczerwinski.pl" target="_blank" rel="noopener noreferrer" className="author-link">Damian Czerwiński</a></p>
                 </div>
-                <div className="theme-toggle">
-                    <button
-                        className="theme-btn"
-                        onClick={toggleTheme}
-                        title={`Switch to ${theme === 'dark' ? 'light' : 'dark'} mode`}
-                    >
-                        <FontAwesomeIcon
-                            icon={theme === 'dark' ? faSun : faMoon}
-                            className="theme-icon"
-                        />
-                        <span className="theme-text">
-                            {theme === 'dark' ? 'Light' : 'Dark'}
-                        </span>
-                    </button>
+                <div className="footer-links">
+                    <a href="/terms-of-service" className="footer-link">Terms of Service</a>
+                    <a href="/privacy-policy" className="footer-link">Privacy Policy</a>
                 </div>
             </div>
         </footer>
