+++ application/src/styles/base.css
@@ -98,7 +98,8 @@ button {
 
 button:disabled {
     cursor: not-allowed;
-    opacity: 0.6;
+    opacity: 0.65;
+    filter: contrast(1.2);
 }
 
 input,
