+++ application/src/styles/floating-action-button.css
@@ -1,6 +1,6 @@
 .floating-action-button {
     position: fixed;
-    bottom: 80px;
+    bottom: calc(80px + var(--spacing-xl));
     right: var(--spacing-xl);
     width: 56px;
     height: 56px;
@@ -13,18 +13,28 @@
     box-shadow: var(--shadow-lg);
     transition: all var(--transition-base);
     z-index: var(--z-overlay);
+    display: -webkit-box;
+    display: -ms-flexbox;
     display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
     align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
     justify-content: center;
 }
 
 .floating-action-button:hover {
     background-color: var(--accent-hover);
+    -webkit-transform: scale(1.1);
+    -ms-transform: scale(1.1);
     transform: scale(1.1);
     box-shadow: var(--shadow-xl);
 }
 
 .floating-action-button:active {
+    -webkit-transform: scale(0.95);
+    -ms-transform: scale(0.95);
     transform: scale(0.95);
 }
 
@@ -35,7 +45,7 @@
 
 @media (max-width: 768px) {
     .floating-action-button {
-        bottom: 70px;
+        bottom: calc(95px + var(--spacing-sm));
         right: var(--spacing-lg);
         width: 52px;
         height: 52px;
@@ -45,7 +55,7 @@
 
 @media (max-width: 480px) {
     .floating-action-button {
-        bottom: 75px;
+        bottom: calc(100px + var(--spacing-sm));
         right: 1.2rem;
         width: 48px;
         height: 48px;
