+++ application/src/styles/note-card.css
@@ -8,8 +8,15 @@
     min-height: 120px;
     word-wrap: break-word;
     box-shadow: var(--shadow-base);
+    display: -webkit-box;
+    display: -ms-flexbox;
     display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
     flex-direction: column;
+    -webkit-box-pack: justify;
+    -ms-flex-pack: justify;
     justify-content: space-between;
 }
 
@@ -19,6 +26,8 @@
 }
 
 .note-content {
+    -webkit-box-flex: 1;
+    -ms-flex: 1;
     flex: 1;
     margin-bottom: var(--spacing-sm);
 }
@@ -44,14 +53,19 @@
 .note-text.completed {
     text-decoration: line-through;
     color: var(--text-secondary);
-    opacity: 0.7;
 }
 
 .note-actions {
+    display: -webkit-box;
+    display: -ms-flexbox;
     display: flex;
+    -webkit-box-pack: end;
+    -ms-flex-pack: end;
     justify-content: flex-end;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
     align-items: center;
-    gap: var(--spacing-xs);
+    gap: var(--spacing-xxs);
     opacity: 0;
     transition: opacity var(--transition-fast);
     margin-top: var(--spacing-sm);
@@ -67,13 +81,18 @@
     padding: var(--spacing-xs);
     border-radius: var(--radius-sm);
     transition: all var(--transition-fast);
-    font-size: 0.75rem;
-    opacity: 0.8;
+    font-size: 0.7rem;
+    display: -webkit-box;
+    display: -ms-flexbox;
     display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
     align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
     justify-content: center;
-    width: 24px;
-    height: 24px;
+    width: 22px;
+    height: 22px;
     border: 1px solid var(--border-color);
     background-color: var(--bg-primary);
 }
@@ -83,6 +102,8 @@
     opacity: 1;
     border-color: var(--text-secondary);
     color: var(--text-primary);
+    -webkit-transform: scale(1.1);
+    -ms-transform: scale(1.1);
     transform: scale(1.1);
 }
